+++ samples/src/main/kotlin/jp/shiguredo/sora/sample/ui/SpotlightRoomSetupActivity.kt
@@ -37,6 +37,7 @@ class SpotlightRoomSetupActivity : AppCompatActivity() {
     private var audioEnabledSpinner: MaterialSpinner? = null
     private var sizeSpinner:         MaterialSpinner? = null
     private var fpsSpinner:          MaterialSpinner? = null
+    private var sdpSemanticsSpinner: MaterialSpinner? = null
 
     val spotlightOptions = listOf(2, 1, 3, 4, 5)
     val videoCodecOptions = listOf("VP8", "VP9", "H264")
@@ -47,6 +48,7 @@ class SpotlightRoomSetupActivity : AppCompatActivity() {
     val bitRateOptions = listOf("UNDEFINED", "100", "300", "500", "800", "1000", "1500", "2000", "2500")
     val sizeOptions = listOf("VGA", "QQVGA", "QCIF", "HQVGA", "QVGA", "HD", "FHD")
     val fpsOptions = listOf("30", "10", "15", "20", "24", "60")
+    val sdpSemanticsOptions = listOf("Plan B", "Unified Plan")
 
     private fun setupUI() {
 
@@ -447,6 +449,44 @@ class SpotlightRoomSetupActivity : AppCompatActivity() {
                         fpsSpinner?.setItems(fpsOptions)
                     }
 
+                    relativeLayout {
+
+                        lparams{
+                            width = matchParent
+                            height= wrapContent
+                            margin = dip(10)
+                        }
+
+                        backgroundColor = Color.parseColor(spinnerBackgroundColor)
+
+                        textView {
+                            maxLines = 10
+                            text = "sdpSemantics"
+                            padding = dip(10)
+                            backgroundColor = Color.parseColor(spinnerBackgroundColor)
+                        }.lparams {
+                            width = wrapContent
+                            height = wrapContent
+                            margin = dip(10)
+                            alignParentLeft()
+                            centerVertically()
+                        }
+
+                        sdpSemanticsSpinner = materialSpinner {
+                            padding = dip(10)
+
+                            lparams{
+                                width = dip(spinnerWidth)
+                                height= wrapContent
+                                margin = dip(10)
+                                alignParentRight()
+                                centerVertically()
+                            }
+                        }
+
+                        sdpSemanticsSpinner?.setItems(sdpSemanticsOptions)
+                    }
+
                 }
 
             }
@@ -470,6 +510,7 @@ class SpotlightRoomSetupActivity : AppCompatActivity() {
         val bitRate = bitRateOptions[bitRateSpinner!!.selectedIndex]
         val size = sizeOptions[sizeSpinner!!.selectedIndex]
         val fps = fpsOptions[fpsSpinner!!.selectedIndex]
+        val sdpSemantics = sdpSemanticsOptions[sdpSemanticsSpinner!!.selectedIndex]
 
         val intent = Intent(this, VideoChatRoomActivity::class.java)
         intent.putExtra("CHANNEL_NAME", channelName)
@@ -482,6 +523,7 @@ class SpotlightRoomSetupActivity : AppCompatActivity() {
         intent.putExtra("BITRATE", bitRate)
         intent.putExtra("VIDEO_SIZE", size)
         intent.putExtra("FPS", fps)
+        intent.putExtra("SDP_SEMANTICS", sdpSemantics)
 
         startActivity(intent)
     }