+++ webrtc-video-effector/src/main/java/jp/shiguredo/webrtc/video/effector/CapturerObserverProxy.java
@@ -2,6 +2,7 @@ package jp.shiguredo.webrtc.video.effector;
 
 import android.os.Handler;
 
+import org.webrtc.CapturerObserver;
 import org.webrtc.NV12Buffer;
 import org.webrtc.SurfaceTextureHelper;
 import org.webrtc.ThreadUtils;
@@ -11,14 +12,14 @@ import org.webrtc.YuvHelper;
 
 import java.nio.ByteBuffer;
 
-public class CapturerObserverProxy implements VideoCapturer.CapturerObserver {
+public class CapturerObserverProxy implements CapturerObserver {
     public static final String TAG = CapturerObserverProxy.class.getSimpleName();
 
-    private VideoCapturer.CapturerObserver originalObserver;
+    private CapturerObserver originalObserver;
     private RTCVideoEffector videoEffector;
 
     public CapturerObserverProxy(final SurfaceTextureHelper surfaceTextureHelper,
-                                 VideoCapturer.CapturerObserver observer,
+                                 CapturerObserver observer,
                                  RTCVideoEffector effector) {
 
         this.originalObserver = observer;